+++ clawness/__init__.py
@@ -3,4 +3,4 @@
 from .core import Clawness, Rule, load_rules
 
 __all__ = ["Clawness", "Rule", "load_rules"]
-__version__ = "1.6.2"
+__version__ = "1.7.0"